requirements.txt
aiogram==3.24.0
python-dotenv==1.0.1
aiohttp==3.10.11
